提交Borg量表答案；获取Borg量表答案；

+++ src/main/java/hci/dky/dao/BorgScaleAnswerPaperMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BorgScaleAnswerPaperMapper {
-    int countByExample(BorgScaleAnswerPaperExample example);
+    long countByExample(BorgScaleAnswerPaperExample example);
 
     int deleteByExample(BorgScaleAnswerPaperExample example);
